+++ src/Printer/SharkPrinter.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Paraunit\Printer;
 
 use Paraunit\Bin\Paraunit;
+use Paraunit\Lifecycle\AbstractEvent;
 use Paraunit\Lifecycle\BeforeEngineStart;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
@@ -22,7 +23,7 @@ class SharkPrinter extends AbstractPrinter implements EventSubscriberInterface
     }
 
     /**
-     * @return array<string, (string|int)[]>
+     * @return array<class-string<AbstractEvent>, array{0: string, 1: int}>
      */
     public static function getSubscribedEvents(): array
     {