+++ src/Process/SymfonyProcessWrapper.php
@@ -12,9 +12,6 @@ class SymfonyProcessWrapper extends AbstractParaunitProcess
     /** @var Process */
     private $process;
 
-    /** @var string */
-    protected $commandLine;
-
     /**
      * {@inheritdoc}
      */
@@ -45,7 +42,7 @@ class SymfonyProcessWrapper extends AbstractParaunitProcess
      */
     public function getOutput(): string
     {
-        return $this->process->getOutput() ?? '';
+        return $this->process->getOutput();
     }
 
     /**
@@ -53,7 +50,7 @@ class SymfonyProcessWrapper extends AbstractParaunitProcess
      */
     public function getErrorOutput(): string
     {
-        return $this->process->getErrorOutput() ?? '';
+        return $this->process->getErrorOutput();
     }
 
     /**