+++ src/Proxy/XDebugProxy.php
@@ -10,4 +10,23 @@ class XDebugProxy
     {
         return extension_loaded('xdebug');
     }
+
+    /**
+     * @psalm-return 2|3
+     */
+    public function getMajorVersion(): int
+    {
+        if (! $this->isLoaded()) {
+            throw new \RuntimeException('Xdebug is not loaded');
+        }
+
+        /** @var string $xdebugVersion */
+        $xdebugVersion = phpversion('xdebug');
+
+        if (version_compare('3.0', $xdebugVersion, '<=')) {
+            return 3;
+        }
+
+        return 2;
+    }
 }